adding config factories
